backend/requirements.txt
# PrintVault Backend

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
sqlalchemy>=2.0.25
python-multipart>=0.0.6
aiofiles>=23.2.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
trimesh>=4.0.0
numpy>=1.26.0
pillow>=10.2.0
websockets>=12.0
httpx>=0.26.0
cryptography>=42.0.0
alembic>=1.13.0
